Feat: refactor types folder
