+++ src/app/api/getFinalityProviders.ts
@@ -1,7 +1,7 @@
 import { encode } from "url-safe-base64";
 import { apiWrapper } from "./apiWrapper";
-import { FinalityProvider } from "../types/finalityProviders";
-import { Pagination } from "../types/api";
+import { FinalityProvider } from "../common/types/finalityProviders";
+import { Pagination } from "./types";
 
 export interface PaginatedFinalityProviders {
   finalityProviders: FinalityProvider[];
