+++ src/app/common/types/errors.ts
