+++ src/app/common/types/globalParams.ts
