+++ src/app/common/types/stakingParams.ts
