+++ src/app/components/Delegations/types.ts
