+++ src/app/components/FinalityProviders/FinalityProviders.tsx
@@ -2,13 +2,13 @@ import { AiOutlineInfoCircle } from "react-icons/ai";
 import InfiniteScroll from "react-infinite-scroll-component";
 import { Tooltip } from "react-tooltip";
 
-import { FinalityProvider as FinalityProviderInterface } from "@/app/types/finalityProviders";
+import { FinalityProvider as FinalityProviderInterface } from "@/app/common/types/finalityProviders";
 import {
   LoadingTableList,
   LoadingView,
 } from "@/app/components/Loading/Loading";
-import { QueryMeta } from "@/app/types/api";
 import { FinalityProvider } from "./FinalityProvider";
+import { QueryMeta } from "@/app/api/types";
 
 interface FinalityProvidersProps {
   data: FinalityProviderInterface[] | undefined;
