+++ src/app/components/Stakers/Stakers.tsx
@@ -6,7 +6,7 @@ import { Tooltip } from "react-tooltip";
 import { LoadingView } from "@/app/components/Loading/Loading";
 import { getStakers } from "@/app/api/getStakers";
 import { useError } from "@/app/context/Error/ErrorContext";
-import { ErrorState } from "@/app/types/errors";
+import { ErrorState } from "@/app/common/types/errors";
 import { Staker } from "./Staker";
 
 interface StakersProps {}
