+++ src/app/components/Stats/types.ts
