+++ src/app/mock/Delegations/mockGenerateDelegationsData.tsx
@@ -1,4 +1,4 @@
-import { Delegation } from "@/app/types/delegations";
+import { Delegation } from "@/app/components/Delegations/types";
 
 function generateRandomHex(size: number) {
   return [...Array(size)]
